requirements.txt
scikit-learn
pickleshare
ninja
gdown
easydict
PyYAML
protobuf
tensorboard
termcolor
tqdm
multimethod
h5py
matplotlib
wandb
pyvista
setuptools
Cython
pandas
deepspeed
opencv-python
seaborn
open3d
mayavi
torch
torchvision
